+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added basic group support. (@daniel-brosche)
 - Added `--force-interactive` option to interactively overwrite changed dependencies on install or update command. (@daniel-brosche)
+- **BREAKING**: Renamed `-f` alias to `-F` (`-f` now implies `--force-interactive`).
 
 # 1.6 (2019-01-26)